requirements.txt
# Core dependencies for SLO detection application
flask==3.0.3
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.0

# Additional dependencies for docx handling
lxml>=4.9.0